+++ NEWS.md
@@ -1,4 +1,17 @@
-# zoomerjoin 0.1.2
+# zoomerjoin 0.1.5
+
+## New features
+
+* Several performance improvements (#101, #104).
+* Added support for joining based on hamming distance (#100).
+* Bumped `extendr` to v0.7.0 (#121)
+
+## Bug fixes
+
+* Fixed bug where when `clean = TRUE`, strings were not coerced to lower case (#105).
+* Fix argument `progress`, was inoperative (#107).
+
+# zoomerjoin 0.1.4
 
 * Submitted Package to CRAN
 * Add support for new `join_by()` syntax